% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cd_competitions.R
\name{fetch_cd_comps}
\alias{fetch_cd_comps}
\title{Fetch Champion Data Competitions List}
\usage{
fetch_cd_comps()
}
\value{
A tibble containing competition details.
}
\description{
Retrieves the list of competitions from Champion Data MC API.
}
