% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_coaches.R
\name{fetch_coaches}
\alias{fetch_coaches}
\title{Fetch Rugby League Coaches (main wrapper)}
\usage{
fetch_coaches(
  season,
  league = c("nrl", "super_league", "championship", "league_one", "womens_super_league",
    "qld_cup", "nsw_cup", "state_of_origin", "challenge_cup", "1895_cup"),
  source = c("rugbyleagueproject")
)
}
\arguments{
\item{season}{Integer. Season year (from 1998).}

\item{league}{Character. One of: "nrl", "super_league", "championship", "league_one",
"womens_super_league", "qld_cup", "nsw_cup",
"state_of_origin", "challenge_cup", "1895_cup".}

\item{source}{Character. Only "rugbyleagueproject" is currently supported.}
}
\value{
A tibble of coach details.
}
\description{
Fetch Rugby League Coaches (main wrapper)
}
\examples{
fetch_coaches(2024, league = "nrl")
}
