% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_lineup.R
\name{fetch_lineups}
\alias{fetch_lineups}
\title{Fetch NRL Team Lineups}
\usage{
fetch_lineups(url, source = "nrl.com", type = "team_list")
}
\arguments{
\item{url}{Character. The full URL to the team list page on nrl.com.}

\item{source}{Character. Currently only "nrl.com" is supported.}

\item{type}{Character. Currently only "team_list".}
}
\value{
A tibble with game, first_name, last_name, team, role.
}
\description{
Fetches NRL team lineups from nrl.com for a given round using a direct URL.
Prints a simple CLI message about the URL being fetched.
}
\examples{
fetch_lineups(url = "https://www.nrl.com/news/2024/05/07/nrl-team-lists-round-10/")
}
