% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_player_stats.R
\name{fetch_player_stats_championdata}
\alias{fetch_player_stats_championdata}
\title{Fetch Champion Data Player Stats for Competition}
\usage{
fetch_player_stats_championdata(comp, round = NULL)
}
\arguments{
\item{comp}{Integer. Competition ID (required).}

\item{round}{Integer or NULL. Round filter (optional).}
}
\value{
Tibble of player stats joined with fixture info.
}
\description{
Fetches player stats for the given Champion Data competition ID. Use fetch_cd_competitions for all ids.
If round is provided, filters fixtures to that round.
Otherwise fetches all completed matches available.
}
